+++ src/main/java/com/commercehub/gradle/plugin/avro/AvroExtension.java
@@ -6,4 +6,5 @@ public interface AvroExtension {
     String getFieldVisibility();
     String getTemplateDirectory();
     boolean isCreateSetters();
+    boolean isRetryDuplicateTypes();
 }