+++ src/main/java/com/commercehub/gradle/plugin/avro/Constants.java
@@ -18,6 +18,7 @@ class Constants {
     static final String DEFAULT_TEMPLATE_DIR = System.getProperty("org.apache.avro.specific.templates",
                                                                   "/org/apache/avro/compiler/specific/templates/java/classic/");
     static final boolean DEFAULT_CREATE_SETTERS = true;
+    static final boolean DEFAULT_RETRY_DUPLICATE_TYPES = true;
 
     static final String SCHEMA_EXTENSION = "avsc";
     static final String PROTOCOL_EXTENSION = "avpr";
@@ -33,4 +34,5 @@ class Constants {
     static final String OPTION_FIELD_VISIBILITY = "fieldVisibility";
     static final String OPTION_STRING_TYPE = "stringType";
     static final String OPTION_ENCODING = "encoding";
+    static final String OPTION_RETRY_DUPLICATE_TYPES = "retryDuplicateTypes";
 }